+++ karma.conf.js
@@ -1,38 +1,27 @@
+/* eslint-env node */
 // Karma configuration
-// http://karma-runner.github.io/0.13/config/configuration-file.html
-/*eslint-env node*/
+// Docs: http://karma-runner.github.io/1.0/config/configuration-file.html
 
 import makeWebpackConfig from './webpack.make';
 
 module.exports = function(config) {
   config.set({
-    // base path, that will be used to resolve files and exclude
-    basePath: '',
+    // Disable watching files
+    autoWatch: false,
 
-    // testing framework to use (jasmine/mocha/qunit/...)
-    frameworks: ['mocha', 'chai', 'sinon-chai', 'chai-as-promised', 'chai-things'],
+    // base path, that will be used to resolve files and exclude
+    //basePath: '', // Commented due to being the default
+
+    browsers: ['PhantomJS'], // PhantomJS, Chrome, ChromeCanary, Firefox, Opera, IE, Safari
 
     client: {
+      // Does not appear to be standard to karma.conf
       mocha: {
         timeout: 5000 // set default mocha spec timeout
       }
     },
 
-    // list of files / patterns to load in the browser
-    files: ['spec.js'],
-
-    preprocessors: {
-      'spec.js': ['webpack']
-    },
-
-    webpack: makeWebpackConfig({ TEST: true }),
-
-    webpackMiddleware: {
-      // webpack-dev-middleware configuration
-      // i. e.
-      noInfo: true
-    },
-
+    // Does not appear to be standard to karma.conf
     coverageReporter: {
       reporters: [{
         type: 'html', //produces a html document after code is run
@@ -45,29 +34,25 @@ module.exports = function(config) {
       dir: 'coverage/' //path to created html doc
     },
 
-    plugins: [
-      require('karma-chrome-launcher'),
-      require('karma-coverage'),
-      require('karma-firefox-launcher'),
-      require('karma-mocha'),
-      require('karma-chai-plugins'),
-
-      require('karma-spec-reporter'),
-      require('karma-phantomjs-launcher'),
-      require('karma-script-launcher'),
-      require('karma-webpack'),
-      require('karma-sourcemap-loader')
-    ],
-
     // list of files / patterns to exclude
-    exclude: [],
+    //exclude: [], // Default commented out
+
+    // list of files / patterns to load in the browser
+    files: ['spec.js'],
+
+    frameworks: ['chai', 'chai-as-promised', 'chai-things', 'mocha', 'sinon-chai'],
+
+    //logLevel: config.LOG_INFO, // LOG_INFO || LOG_DISABLE || LOG_ERROR || LOG_WARN || LOG_INFO || LOG_DEBUG // Default commented out
+
+    // List: https://www.npmjs.com/browse/keyword/karma-plugin
+    //plugins: ['karma-*'], // Default commented out
 
     // web server port
     port: 9000,
 
-    // level of logging
-    // possible values: LOG_DISABLE || LOG_ERROR || LOG_WARN || LOG_INFO || LOG_DEBUG
-    logLevel: config.LOG_INFO,
+    preprocessors: {
+      'spec.js': ['webpack', 'sourcemap'] // added karma-sourcemap-loader today
+    },
 
     // reporter types:
     // - dots
@@ -76,23 +61,15 @@ module.exports = function(config) {
     // - junit
     // - growl
     // - coverage
-    reporters: ['spec', 'coverage'],
+    reporters: ['spec', 'coverage'], // should 'spec' be 'mocha'?
 
-    // enable / disable watching file and executing tests whenever any file changes
-    autoWatch: false,
+    // Continuous Integration mode - if true, capture browsers, run tests and exit
+    //singleRun: false, // Commented due to being the default
 
-    // Start these browsers, currently available:
-    // - Chrome
-    // - ChromeCanary
-    // - Firefox
-    // - Opera
-    // - Safari (only Mac)
-    // - PhantomJS
-    // - IE (only Windows)
-    browsers: ['PhantomJS'],
+    webpack: makeWebpackConfig({ TEST: true }),
 
-    // Continuous Integration mode
-    // if true, it capture browsers, run tests and exit
-    singleRun: false
+    webpackMiddleware: {
+      noInfo: true
+    }
   });
 };