+++ webpack.make.js
@@ -7,33 +7,38 @@ import HtmlWebpackPlugin from 'html-webpack-plugin';
 import HtmlWebpackHarddiskPlugin from 'html-webpack-harddisk-plugin';
 import ExtractTextPlugin from 'extract-text-webpack-plugin';
 import path from 'path';
-import util from 'util';
 
 module.exports = function makeWebpackConfig(options) {
-  /**
-   * Environment type
-   * BUILD is for generating minified builds
-   * TEST is for generating test builds
-  **/
-  let BUILD = !!options.BUILD;
-  let TEST = !!options.TEST;
-  let E2E = !!options.E2E;
-  let DEV = !!options.DEV;
+  // Set by gulpfile.babel.js or karma.conf.js
+  const BUILD = !!options.BUILD;
+  const TEST = !!options.TEST;
+  const E2E = !!options.E2E;
+  const DEV = !!options.DEV;
 
-  let CommonsChunkPlugin = webpack.optimize.CommonsChunkPlugin;
+  let config = {
+    cache: DEV,
 
-  let config = {};
+    //context: __dirname, // current directory by default
 
-  if(TEST) {
-    config.entry = {}; // empty for test builds (Karma will handle)
-  } else {
-    config.entry = {
+    devServer: {
+      contentBase: './client/',
+      stats: {
+        modules: false,
+        cached: false,
+        chunks: false
+      }
+    },
+
+    devtool: '', // placeholder to be filled in conditionally
+
+    // If testing, must set entry to '' to avoid Karma error (bug)
+    entry: TEST ? '' : {
       app: './client/app/app.js',
       polyfills: './client/polyfills.js',
       vendor: [
         'angular',
         'angular-aria',
-        // 'angular-animate',
+        // 'angular-animate', // Not using ngAnimate
         'angular-cookies',
         'angular-resource',
         'angular-messages',
@@ -42,43 +47,103 @@ module.exports = function makeWebpackConfig(options) {
         'angular-ui-router',
         'lodash'
       ]
-    };
-  }
+    },
 
-  if(TEST) {
-    config.output = {}; // empty for test builds (Karma will handle)
-  } else {
-    config.output = {
-      // Absolute output directory
-      path: BUILD ? path.join(__dirname, '/dist/client/') : path.join(__dirname, '/.tmp/'),
+    module: {
+      rules: [
+        {
+          test: /\.js$/,
+          loader: 'babel-loader', // transpiles ES6-7 to ES5
+          exclude: /node_modules/,
+          options: {
+            //babelrc: false, // ignore babel settings in babelrc and package.json
+            //presets: [['es2015', {modules: false}]],
+            cacheDirectory: true,
+            shouldPrintComment: commentContents => /@ngInject/.test(commentContents), // leave ng-annotate alone
+            plugins: TEST ? ['istanbul', 'transform-class-properties'] : ['transform-class-properties']
+          },
+          include: [
+            path.resolve(__dirname, 'client/'),
+            path.resolve(__dirname, 'node_modules/lodash-es/')
+          ]
+        },
 
-      // Output path from the view of the page
-      // Uses webpack-dev-server in development
-      publicPath: BUILD || DEV || E2E ? '/' : `http://localhost:${8080}/`,
-      //publicPath: BUILD ? '/' : `http://localhost:${env.port}/`,
+        {
+          test: /\.(png|jpg|jpeg|gif|svg|woff|woff2|ttf|eot)([?]?.*)$/,
+          loader: 'file-loader' // copies to dist, renaming with asset hash
+        },
 
-      // Filename for entry points
-      // Only adds hash in build mode
-      filename: BUILD ? '[name].[hash].js' : '[name].bundle.js',
+        {
+          test: /\.pug$/,
+          loader: 'pug-html-loader' // converts pug to HTML (includes pug node module)
+        },
 
-      // Filename for non-entry points
-      // Only adds hash in build mode
-      chunkFilename: BUILD ? '[name].[hash].js' : '[name].bundle.js'
-    };
-  }
+        {
+          test: /\.css$/,
+          use: !TEST
+            // Docs: https://github.com/webpack/style-loader
+            // Use style-loader in development for hot-loading
+            ? ExtractTextPlugin.extract({
+              fallback: 'style-loader',
+              use: [
+                {
+                  loader: 'css-loader'
+                },
+                {
+                  loader: 'postcss-loader',
+                  options: {
+                    plugins: () => [autoprefixer({browsers: ['last 2 versions']})]
+                  }
+                }
+              ]
+            })
+            // Docs: https://github.com/webpack/null-loader
+            // Skip loading css in test mode
+            : [{loader: 'null-loader'}]
+        },
 
-  if(TEST) {
-    config.resolve = {
-      modulesDirectories: ['node_modules'],
-      extensions: ['', '.js']
-    };
-  }
+        {
+          test: /\.scss$/,
+          use: [
+            {loader: 'style-loader'},
+            {loader: 'css-loader'},
+            {
+              loader: 'sass-loader',
+              options: {
+                outputStyle: 'compressed',
+                precision: 10,
+                sourceComments: false
+              }
+            }
+          ],
+          include: [
+            path.resolve(__dirname, 'node_modules/bootstrap-sass/assets/stylesheets/*.scss'),
+            path.resolve(__dirname, 'client/app/app.scss')
+          ]
+        },
 
-/**
- * Devtool
- * Reference: http://webpack.github.io/docs/configuration.html#devtool
- * Type of sourcemap to use per build type
- */
+        {
+          // Explore https://github.com/jeffling/ng-annotate-webpack-plugin as
+          // possible replacement for ng-annotate-loader (seemed slow though)
+          test: /\.js$/,
+          loader: 'ng-annotate-loader?single_quotes',
+          enforce: 'post'
+        }
+      ]
+    },
+
+    node: {
+      crypto: 'empty',
+      clearImmediate: false, // do not stop timer associated with callback
+      setImmediate: false // do not schedule immediate callback
+    },
+
+    output: {}, // placeholder to be filled in conditionally
+
+    plugins: [] // placeholder to be filled in conditionally
+  };
+
+  // Type of sourcemap to use per build type
   if(TEST) {
     config.devtool = 'inline-source-map';
   } else if(BUILD || DEV) {
@@ -87,146 +152,30 @@ module.exports = function makeWebpackConfig(options) {
     config.devtool = 'eval';
   }
 
-/**
- * Loaders
- * Reference: http://webpack.github.io/docs/configuration.html#module-loaders
- * List: http://webpack.github.io/docs/list-of-loaders.html
- * This handles most of the magic responsible for converting modules
- */
-
-  // Moves to new webpack.LoaderOptionsPlugin in Webpack 2
-  config.sassLoader = {
-    outputStyle: 'compressed',
-    precision: 10,
-    sourceComments: false
-  };
-
-  // Moves to new webpack.LoaderOptionsPlugin in Webpack 2
-  config.babel = {
-    shouldPrintComment(commentContents) {
-      // keep `/*@ngInject*/`
-      return /@ngInject/.test(commentContents);
-    }
-  };
-
-  // Initialize module
-  config.module = {
-    preLoaders: [],
-    loaders: [
-      {
-        // JS LOADER
-        // Reference: https://github.com/babel/babel-loader
-        // Transpile .js files using babel-loader
-        // Compiles ES6 and ES7 into ES5 code
-        test: /\.js$/,
-        loader: 'babel-loader',
-        exclude: /node_modules/,
-        include: [
-          path.resolve(__dirname, 'client/'),
-          path.resolve(__dirname, 'node_modules/lodash-es/')
-        ]
-      },
-      {
-        // ASSET LOADER
-        // Reference: https://github.com/webpack/file-loader
-        // Copy png, jpg, jpeg, gif, svg, woff, woff2, ttf, eot files to output
-        // Rename the file using the asset hash
-        // Pass along the updated reference to your code
-        // You can add here any file extension you want to get copied to your output
-        test: /\.(png|jpg|jpeg|gif|svg|woff|woff2|ttf|eot)([?]?.*)$/,
-        loader: 'file-loader'
-      },
-      {
-        // JSON LOADER
-        test: /\.json$/,
-        loader: 'json-loader'
-      },
-      {
-        // Pug HTML LOADER
-        // Reference: https://github.com/willyelm/pug-html-loader
-        // Allow loading Pug throw js
-        test: /\.pug$/,
-        loader: 'pug-html-loader'
-      },
-      {
-        // CSS LOADER
-        // Reference: https://github.com/webpack/css-loader
-        // Allow loading css through js
-        //
-        // Reference: https://github.com/postcss/postcss-loader
-        // Postprocess your css with PostCSS plugins
-        test: /\.css$/,
-        loader: !TEST
-          // Reference: https://github.com/webpack/extract-text-webpack-plugin
-          // Extract css files in production builds
-          //
-          // Reference: https://github.com/webpack/style-loader
-          // Use style-loader in development for hot-loading
-          ? ExtractTextPlugin.extract('style-loader', 'css-loader!postcss-loader')
-          // Reference: https://github.com/webpack/null-loader
-          // Skip loading css in test mode
-          : 'null-loader'
-      },
-      {
-        // SASS LOADER
-        // Reference: https://github.com/jtangelder/sass-loader
-        test: /\.scss$/,
-        loaders: ['style-loader', 'css-loader', 'sass-loader'],
-        include: [
-          path.resolve(__dirname, 'node_modules/bootstrap-sass/assets/stylesheets/*.scss'),
-          path.resolve(__dirname, 'client/app/app.scss')
-        ]
-      }]
-  };
-
-  config.module.postLoaders = [{
-    test: /\.js$/,
-    loader: 'ng-annotate-loader?single_quotes'
-  }];
-
-  // ISPARTA LOADER (instanbul-instrumenter-loader has issues)
-  // Reference: https://github.com/deepsweet/istanbul-instrumenter-loader
-  // Instrument JS files with Isparta (as set in Gulpfile) for subsequent code coverage reporting
-  // Skips node_modules and spec files
-  if(TEST) {
-    config.module.preLoaders.push({
-      //delays coverage til after tests are run, fixing transpiled source coverage error
-      test: /\.js$/,
-      exclude: /(node_modules|spec\.js|mock\.js)/,
-      loader: 'isparta-loader',
-      query: {
-        babel: {
-          // optional: ['runtime', 'es7.classProperties', 'es7.decorators']
-        }
-      }
-    });
-  }
-
-  // Moves to new webpack.LoaderOptionsPlugin in Webpack 2
-  config.postcss = [
-    autoprefixer({
-      browsers: ['last 2 versions']
-    })
-  ];
-
-  config.plugins = [
-    // Separate CSS https://github.com/webpack/extract-text-webpack-plugin
-    // Disabled when in test mode or not in build mode
-    new ExtractTextPlugin('[name].[hash].css', {
-      disable: !BUILD || TEST
-    })
-  ];
-
   if(!TEST) {
-    config.plugins.push(new CommonsChunkPlugin({
+    config.output = {
+      // Absolute output directory
+      path: BUILD ? path.join(__dirname, '/dist/client/') : path.join(__dirname, '/.tmp/'),
+
+      // Output path from the view of the page - uses webpack-dev-server in development
+      publicPath: BUILD || DEV || E2E ? '/' : `http://localhost:${8080}/`,
+      //publicPath: BUILD ? '/' : `http://localhost:${env.port}/`,
+
+      // Filename for entry points - add hash in build mode
+      filename: BUILD ? '[name].[hash].js' : '[name].bundle.js',
+
+      // Filename for non-entry points - add hash in build mode
+      chunkFilename: BUILD ? '[name].[hash].js' : '[name].bundle.js'
+    };
+
+    config.plugins.push(new webpack.optimize.CommonsChunkPlugin({
       name: 'vendor', // vendor.js
       minChunks: Infinity // chunk is only for vendor JS
     }));
-  }
 
-  if(!TEST) {
     config.plugins.push(
-      // Don't render index.html (https://github.com/ampedandwired/html-webpack-plugin)
+      // Docs: https://github.com/ampedandwired/html-webpack-plugin
+      // Don't render index.html
       new HtmlWebpackPlugin({
         template: 'client/_index.html',
         filename: '../client/index.html',
@@ -236,18 +185,30 @@ module.exports = function makeWebpackConfig(options) {
     );
   }
 
+  if(DEV) {
+    config.plugins.push(
+      // Docs: https://webpack.github.io/docs/list-of-plugins.html#defineplugin
+      // Define free global variables
+      new webpack.DefinePlugin({
+        'process.env': {
+          NODE_ENV: '"development"'
+        }
+      })
+    );
+    // config.plugins.push(
+    //   new webpack.HotModule.ReplacementPlugin()
+    // );
+  }
+
   // Add build specific plugins
   if(BUILD) {
     config.plugins.push(
-      // Reference: http://webpack.github.io/docs/list-of-plugins.html#noerrorsplugin
+      new ExtractTextPlugin({filename: '[name].[hash].css'}),
+      // Docs: http://webpack.github.io/docs/list-of-plugins.html#noerrorsplugin
       // Only emit files when there are no errors
-      new webpack.NoErrorsPlugin(),
+      new webpack.NoEmitOnErrorsPlugin(),
 
-      // Reference: http://webpack.github.io/docs/list-of-plugins.html#dedupeplugin
-      // Dedupe modules in the output
-      new webpack.optimize.DedupePlugin(),
-
-      // Reference: http://webpack.github.io/docs/list-of-plugins.html#uglifyjsplugin
+      // Docs: http://webpack.github.io/docs/list-of-plugins.html#uglifyjsplugin
       // Minify all javascript, switch loaders to minimizing mode
       new webpack.optimize.UglifyJsPlugin({
         mangle: false,
@@ -255,12 +216,12 @@ module.exports = function makeWebpackConfig(options) {
         output: {
           comments: false
         },
-        compress: {
-          warnings: false
-        }
+        // compress: { // not needed as it's a default
+        //   warnings: false
+        // }
       }),
 
-      // Reference: https://webpack.github.io/docs/list-of-plugins.html#defineplugin
+      // Docs: https://webpack.github.io/docs/list-of-plugins.html#defineplugin
       // Define free global variables
       new webpack.DefinePlugin({
         'process.env': {
@@ -270,51 +231,15 @@ module.exports = function makeWebpackConfig(options) {
     );
   }
 
-  if(DEV) {
-    config.plugins.push(
-      // Reference: https://webpack.github.io/docs/list-of-plugins.html#defineplugin
-      // Define free global variables
-      new webpack.DefinePlugin({
-        'process.env': {
-          NODE_ENV: '"development"'
-        }
-      })
-    );
-  }
-
-  config.cache = DEV;
-
-  if(TEST) {
-    config.stats = {
-      colors: true,
-      reasons: true
-    };
-    config.debug = false;
-  }
-
-  /**
-   * Dev server configuration
-   * Reference: http://webpack.github.io/docs/configuration.html#devserver
-   * Reference: http://webpack.github.io/docs/webpack-dev-server.html
-   */
-  config.devServer = {
-    contentBase: './client/',
-    stats: {
-      modules: false,
-      cached: false,
-      colors: true,
-      chunk: false
-    }
-  };
-
-  config.node = {
-    crypto: 'empty',
-    clearImmediate: false, // do not stop timer associated with callback
-    setImmediate: false // do not schedule immediate callback
-  };
-
-  // Show the full Webpack config object
-  console.log('Webpack Configuration:', util.inspect(config, { showHidden: false, depth: null }));
-
+  // For debugging
+  // const util = require('util');
+  //console.log('Webpack Config (as entered):', util.inspect(config, { showHidden: false, depth: null }));
+  // var compiler;
+  // try {
+  //   compiler = webpack(config);
+  //   console.log('Webpack Config (processed):', util.inspect(compiler.options, { showHidden: false, depth: null }));
+  // } catch(e) {
+  //   console.error(e.message);
+  // }
   return config;
 };